requirements.txt
kafka-python==2.0.2
pandas==2.2.3
pymongo==4.10.1